# docs/logging/logging_api_spec.yaml
openapi: 3.0.3
info:
  title: AM Centralized Logging API
  description: Standardized logging interface for Python, Java, and Flutter microservices.
  version: 1.0.0
servers:
  - url: http://am-logging-svc/v1
    description: Internal Kubernetes Service

paths:
  /logs:
    post:
      summary: Ingest a log entry
      operationId: ingestLog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogEntry'
      responses:
        '202':
          description: Log accepted for asynchronous processing
        '400':
          description: Invalid log format

  /telemetry/events:
    post:
      summary: Ingest Flutter product / RUM events
      operationId: ingestTelemetry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              properties:
                events:
                  type: array
                  maxItems: 100
                  items:
                    type: object
                    required: [event]
                    properties:
                      event:
                        type: string
                        enum: [screen_view, api_timing, boot_rum, feature_action, session_start, auth_logout]
                      anon_id: { type: string }
                      session_id: { type: string }
                      user_id: { type: string }
                      section: { type: string }
                      screen: { type: string }
                      duration_ms: { type: number }
                      props: { type: object }
      responses:
        '202':
          description: Events accepted and pushed to Loki
        '422':
          description: Invalid batch

components:
  schemas:
    LogEntry:
      type: object
      required:
        - trace_id
        - span_id
        - service
        - timestamp
        - log_type
        - level
        - payload
      properties:
        trace_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        span_id:
          type: string
          example: "span-abc-123"
        service:
          type: string
          example: "am-auth"
        timestamp:
          type: string
          format: date-time
        log_type:
          type: string
          enum: [TECHNICAL, BUSINESS, AUDIT]
        level:
          type: string
          enum: [DEBUG, INFO, WARN, ERROR, CRITICAL]
        context:
          $ref: '#/components/schemas/LogContext'
        payload:
          type: object
          description: Structured log data or message
        exception:
          $ref: '#/components/schemas/ExceptionInfo'
        metadata:
          type: object
          additionalProperties:
            type: string

    LogContext:
      type: object
      properties:
        class:
          type: string
        method:
          type: string
        inputs:
          type: object
        outputs:
          type: object
        latency_ms:
          type: number
          format: float

    ExceptionInfo:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
        stack:
          type: string